Implementation
I haven't tested the demon level but it seems to be fine. I noticed some of the level scenes were repeats, but I just linked them in alphabetical order.

To add more levels moving forward, make the scene, attach the "level_base.gd" script on it, and then add the level to the level_dict found in last_level_updater.gd